Move the GL exception stuff into its own module.
