+++ src/GL/Shader/Fragment.hs
@@ -6,38 +6,22 @@ import Control.Monad
 import Data.Foldable (for_)
 import Data.List (intercalate)
 import Data.Monoid
-import Data.Typeable
 import Foreign.C.String
 import Foreign.Marshal.Alloc
 import Foreign.Ptr
 import Foreign.Storable
-import GHC.Stack
+import GL.Exception
 import Graphics.GL.Core41
 import Graphics.GL.Types
 import Graphics.Shader.Fragment
 import Linear.V3
 import Prelude hiding (IO)
-import qualified System.IO as IO
-
-type IO a = HasCallStack => IO.IO a
 
 newtype Shader = Shader { unShader :: GLuint }
 
 newtype Program = Program { unProgram :: GLuint }
 
 newtype VAO = VAO { unVAO :: GLuint }
-
-data GLError
-  = InvalidEnum
-  | InvalidValue
-  | InvalidOperation
-  | InvalidFramebufferOperation
-  | OutOfMemory
-  | Other String
-  deriving Show
-
-data GLException = GLException GLError CallStack
-  deriving (Typeable)
 
 toGLSL :: Fragment () -> String
 toGLSL shader
@@ -117,34 +101,3 @@ checkShader = fmap Shader . checkStatus glGetShaderiv glGetShaderInfoLog GL_COMP
 
 checkProgram :: Program -> IO Program
 checkProgram = fmap Program . checkStatus glGetProgramiv glGetProgramInfoLog GL_LINK_STATUS . unProgram
-
-checkStatus :: (GLenum -> GLuint -> Ptr GLint -> IO ()) -> (GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> IO ()) -> GLenum -> GLuint -> IO GLuint
-checkStatus get getLog status object = do
-  success <- alloca $ \ p -> do
-    get object status p
-    peek p
-  when (success == GL_FALSE) $ do
-    l <- alloca $ \ p -> do
-      get object GL_INFO_LOG_LENGTH p
-      peek p
-    log <- allocaBytes (fromIntegral l) $ \ bytes -> do
-      getLog object l nullPtr bytes
-      peekCString bytes
-    throw $ GLException (Other log) callStack
-  pure object
-
-checkGLError :: IO ()
-checkGLError = glGetError >>= \ e -> case e of
-  GL_NO_ERROR -> pure ()
-  GL_INVALID_ENUM -> throw $ GLException InvalidEnum callStack
-  GL_INVALID_VALUE -> throw $ GLException InvalidValue callStack
-  GL_INVALID_OPERATION -> throw $ GLException InvalidOperation callStack
-  GL_INVALID_FRAMEBUFFER_OPERATION -> throw $ GLException InvalidFramebufferOperation callStack
-  GL_OUT_OF_MEMORY -> throw $ GLException OutOfMemory callStack
-  _ -> throw $ GLException (Other "Unknown") callStack
-
-
-instance Show GLException where
-  showsPrec p (GLException e s) = showString "GLException " . showsPrec p e . showChar '\n' . showString (prettyCallStack s)
-
-instance Exception GLException
