+++ ui-effects.cabal
@@ -17,6 +17,7 @@ library
   hs-source-dirs:      src
   exposed-modules:     Control.Applicative.Free.Freer
                      , Control.Monad.Free.Freer
+                     , GL.Exception
                      , GL.Shader
                      , GL.Shader.Fragment
                      , Graphics.Shader.Fragment
